+++ client/src/components/Home/PodcastCard.jsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
+import togglePodcast from '../SubscribedPodcast';
+import User from '../User/User';
 
 // import SubscribeBtn from '../Buttons/SubscribeBtn';
 
@@ -24,7 +26,7 @@ const PodcastImageLink = styled(Link)`
 const PodcastImage = styled.img`
   /* styles for the podcast image */
   max-width: 100;
-  max-height: 150px
+  max-height: 150px;
 `;
 
 const PodcastDetails = styled.div`
@@ -32,11 +34,26 @@ const PodcastDetails = styled.div`
 `;
 
 
-const PodcastCard = ({ podcast, togglePodcast }) => {
+const PodcastCard = ({podcast, user, setUser}) => {
   const { name, itunesInfo, _id } = podcast;
 
-  const handleSubscribeClick = () => {
-    togglePodcast(podcast);
+
+  const isSubscribed = user.subscribedPodcasts.some(id => podcast._id == id)
+
+    const togglePodcast = () => {
+      const updatedPodcasts = 
+      {    
+          "subscribedPodcasts": !isSubscribed ? [...user.subscribedPodcasts, podcast._id ] : user.subscribedPodcasts.filter(id => podcast._id !=id )
+      }
+      const requestOptions = {
+          method: 'PUT',
+          headers: { 'Content-Type': 'application/json' },
+          body: JSON.stringify(updatedPodcasts)
+      }
+      fetch (`http://localhost:9000/api/users/${user._id}`,requestOptions).then(res => res.json()).then (data => {
+            console.log(data)
+            setUser(data)
+        });
   };
 
   return (
@@ -49,8 +66,9 @@ const PodcastCard = ({ podcast, togglePodcast }) => {
       <PodcastDetails>
         <h3>{name}</h3>
         <h5>{podcast.description}</h5>
-        <button onClick={handleSubscribeClick}>
-          {podcast.subscribed ? 'Unsubscribe' : 'Subscribe'}
+        <h4>{isSubscribed}</h4>
+        <button onClick={togglePodcast} >
+          {isSubscribed ? 'Unsubscribe' : 'Subscribe'}
         </button>
       </PodcastDetails>
     </PodcastCardContainer>
